Api clean up
